Refactor aws-lc-fips-sys CMake build logic
